### 2024-02-21 #6387 @dfm
@AnkitBarik — Thanks for your submission! Given the relatively small size of the code base, I have submitted this pre-review for editorial review with respect to the ["substantial scholarly effort"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion. This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.

In the meantime, you could double check the syntax of the metadata in your paper markdown. It looks right to me, and there might be some other bug causing the compilation error above. But, perhaps you can check that it compiles using the tips here: https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles
