### 2024-02-21 #6387 @AnkitBarik
@dfm Thanks a lot! The paper compiles well with the docker command, I'm not sure of the error though. Would it help if I uploaded the PDF to the repository?
