### 2024-02-22 #6387 @xuanxu
The paper does not compile due to the format of the date in the metadata.
[This line](https://github.com/AnkitBarik/planetMagFields/blob/main/paper/paper.md?plain=1#L17) should read: `date: 14 February 2024`
