### 2024-02-22 c21f86b
Fix paper date format
