### 2024-02-22 #6387 @AnkitBarik
Hi, I'm unsure of what the error is, the [open journals github workflow](https://github.com/marketplace/actions/open-journals-pdf-generator) to generate pdf seems to be working on my repository:

https://github.com/AnkitBarik/planetMagFields/actions/runs/8008584107/job/21875333313
