### 2024-02-22 #6387 @dfm
I believe that the affiliation id's should be `1` or `"1"`, not `[1]`.
