### 2024-02-22 4883bf0
Fix ads abbreviations in bib file
