+++ paper/paper.bib
@@ -91,7 +91,7 @@ author = {Thébault, E.  and Langlais, B. and Oliveira, J.S. and Amit, H. and Le
 @article{Connerney1998,
        author = {{Connerney}, J.~E.~P. and {Acu{\~n}a}, M.~H. and {Ness}, N.~F. and {Satoh}, T.},
         title = "{New models of Jupiter's magnetic field constrained by the Io flux tube footprint}",
-      journal = {\jgr},
+      journal = {Journal of Geophysical Research},
      keywords = {Magnetospheric Physics: Planetary magnetospheres, Planetology: Fluid Planets: Magnetic fields and magnetism, Planetology: Fluid Planets: Magnetospheres, Planetology: Solar System Objects: Jupiter},
          year = 1998,
         month = jun,
@@ -315,7 +315,7 @@ year = {2012}
 @article{Wilson2023,
        author = {{Wilson}, R.~J. and {Vogt}, M.~F. and {Provan}, G. and {Kamran}, A. and {James}, M.~K. and {Brennan}, M. and {Cowley}, S.~W.~H.},
         title = "{Internal and External Jovian Magnetic Fields: Community Code to Serve the Magnetospheres of the Outer Planets Community}",
-      journal = {\ssr},
+      journal = {Space Science Reviews},
      keywords = {Jupiter, Magnetic field, Spherical harmonic model, Current sheet model, Community code},
          year = 2023,
         month = feb,
