+++ _sources/api.rst.txt
@@ -1,0 +1,26 @@
+.. planetMagFields documentation master file, created by
+   sphinx-quickstart on Mon Jan 22 08:32:05 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Python API
+=============
+
+.. automodule:: planetmagfields
+   :members:
+
+
+.. autoclass:: planetmagfields.Planet
+   :members:
+   :class-doc-from: class
+
+.. autofunction:: extrapot
+
+.. autofunction:: get_models
+
+.. Indices and tables
+.. ==================
+
+.. * :ref:`genindex`
+.. * :ref:`modindex`
+.. * :ref:`search`
