+++ _sources/examples.rst.txt
@@ -1,0 +1,320 @@
+.. planetMagFields documentation master file, created by
+   sphinx-quickstart on Mon Jan 22 08:32:05 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+.. _secExamples:
+
+#########################
+Features and examples
+#########################
+
+The :py:class:`Planet <planetmagfields.Planet>` class
+*****************************************************
+
+This gives access to all the relevant properties of a planet and has methods to plot
+the field and write a `vts` file for 3D visualization. Usage:
+
+.. code-block:: python
+
+   from planetmagfields import Planet
+   p = Planet(name='earth',datDir='planetmagfields/data/')
+
+
+This displays the some information about the planet
+
+.. code-block:: bash
+
+   Planet: Earth
+   l_max = 13
+   Dipole tilt (degrees) = -9.410531
+
+
+and gives access to
+variables associated with the planet such as:
+
+  * ``p.lmax`` : maximum spherical harmonic degree till which data is available
+  * ``p.glm``, ``p.hlm``: the Gauss coefficients
+  * ``p.Br`` : computed radial magnetic field at surface
+  * ``p.dipTheta`` : dipole tilt with respect to the rotation axis
+  * ``p.dipPhi`` : dipole longitude ( in case zero longitude is known, applicable to Earth )
+  * ``p.idx`` : indices to get values of Gauss coefficients
+  * ``p.model`` : the :ref:`magnetic field model <secmodels>` used. Available models can be obtained using the :py:func:`get_models <planetmagfields.get_models>` function. Selects the latest available model when unspecified.
+
+Example using ``IPython``:
+
+.. code-block:: python
+
+   In [1]: from planetmagfields import Planet
+
+   In [2]: p = Planet(name='jupiter',model='jrm09')
+   Planet: Jupiter
+   Model: jrm09
+   l_max = 10
+   Dipole tilt (degrees) = 10.307870
+
+   In [3]: p.glm[p.idx[2,0]]      # g20
+   Out[3]: 11670.4
+
+   In [4]: p.hlm[p.idx[4,2]]      # h42
+   Out[4]: 27811.2
+
+as well as the functions:
+
+:py:func:`Planet.plot() <planetmagfields.Planet.plot>`
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This function plots a 2D surface plot of the radial magnetic field at radius `r` given in terms of the surface radius.
+For example,
+
+.. code-block:: python
+
+   from planetmagfields import Planet
+   p = Planet(name='earth',datDir='planetmagfields/data/')
+   p.plot(r=1,proj='Mollweide')
+
+
+produces the info mentioned above first and then the following plot of Earth's magnetic field using a Mollweide projection
+
+.. image:: _static/images/2d/earth2d.png
+   :width: 400
+   :align: center
+
+while
+
+.. code-block:: python
+
+   from planetmagfields import Planet
+   p = Planet(name='jupiter',model='jrm09',datDir='planetmagfields/data/')
+   p.plot(r=0.85,proj='Mollweide')
+
+produces the following info about Jupiter and then plot that follows
+
+.. code-block:: bash
+
+   Planet: Jupiter
+   l_max = 10
+   Dipole tilt (degrees) = 10.307870
+
+
+.. image:: _static/images/jupiter_r085.png
+   :width: 400
+   :align: center
+
+This can be compared with Fig. 1 g from `Moore et al. 2018 <https://doi.org/10.1038/s41586-018-0468-5>`_ .
+
[…]
