### 2024-04-23 #6387 @dfm
@steo85it — Thanks!

@kjg136 — Would you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
