### 2024-04-24 #6677 @editorialbot
[REVIEW]: planetMagFields: A Python package for analyzing and plotting planetary magnetic field data
**Submitting author:** @AnkitBarik (<a href="http://orcid.org/0000-0001-5747-669X">Ankit Barik</a>)
**Repository:** https://github.com/AnkitBarik/planetMagFields
**Branch with paper.md** (empty if default branch):
**Version:** 1.4.3
