### 2024-04-29 AnkitBarik/planetMagFields#8 @athulpg007
Non-existent data dir path in basic example for Planet() when using pip
The [installation](https://ankitbarik.github.io/planetMagFields/installation.html) docs state that planetMagFields can be installed from PyPI using `pip install planetMagFields`. The basic example use of class [Planet](https://ankitbarik.github.io/planetMagFields/examples.html#the-planet-class) as stated is
```
from planetmagfields import Planet
p = Planet(name='earth',datDir='planetmagfields/data/')
```
Because the repository was not cloned, the planetmagfields/data directory does not exist, and fails with the error
```
FileNotFoundError: planetmagfields/data/earth_igrf13.dat not found.
```
The issue can be solved by leaving the datDir parameter empty, so it automatically finds the datDir
```
stdDatDir = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'data/')
```
To improve the user experience for first time users who want to `pip install planetMagFields` and just want to try out a basic example, please consider handling this error. One solution is to state before the example that if pip install was used, datDir can be omitted.
