### 2024-04-29 AnkitBarik/planetMagFields#9 @athulpg007
Add Community guidelines for contributions by third parties
Please add a section to the repo readme file on Community guidelines explaining How to: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Please see this example here: https://github.com/astropy/astropy?tab=readme-ov-file#contributing-code-documentation-or-feedback
