### 2024-04-29 AnkitBarik/planetMagFields#10 @athulpg007
Suggested improvements for automated unit tests
Please consider these suggested improvements to the tests:
1. Rename the top level folder to tests/, the test files as test_abc.py, test_xyz.py (starting with "test").
2. Rename the test class as TestABC, TestXYZ etc.
3. [optional] Instead of a single test, consider having more small tests which test specific functions/functionality of the class. It is suggested to include some more parameter combinations for the Planet.__init() class in the tests. There are no tests for other functions of the Planet class, consider adding some tests to verify their functionality.
4. Use pytest to collect and run all the tests in the tests/ directory.
5. [optional] consider adding a github action which runs the pytest step on a PR, and merge to master. This will ensure the basic functionality of the package works, when changes are made.
