### 2024-04-29 #6677 @athulpg007
@AnkitBarik please take a look at the comments below. More details are available in the linked issues.
1. Several example snippets use datDir='planetmagfields/data/' which will not exist if the package was installed using pip. Please see https://github.com/AnkitBarik/planetMagFields/issues/8.
2. Please add a section for "Contributions/Community Guidelines" to README. See https://github.com/AnkitBarik/planetMagFields/issues/9
3. Please see the suggested improvements for automated unit tests. https://github.com/AnkitBarik/planetMagFields/issues/10
