### 2024-04-29 e58ba4f
Update doc for empty data directory
