### 2024-04-29 AnkitBarik/planetMagFields#8 @AnkitBarik
Hi, thank you for pointing this out. This was an issue with the documentation, it has been fixed and now it works as intended:

https://ankitbarik.github.io/planetMagFields/examples.html
