+++ .github/workflows/main.yml
@@ -33,9 +33,9 @@ jobs:
       - name: Install pre-requisites
         run: |
           sudo apt-get update
-          sudo apt-get -y install python3-scipy python3-numpy python3-matplotlib
+          sudo apt-get -y install python3-scipy python3-numpy python3-matplotlib python3-pytest
 
       - name: Run test
         run: |
-          cd ${{github.workspace}}/test
-          python3 planetmagtest.py
+          cd ${{github.workspace}}/tests
+          pytest
