+++ .github/workflows/main.yml
@@ -38,4 +38,4 @@ jobs:
       - name: Run test
         run: |
           cd ${{github.workspace}}/tests
-          pytest
+          python3 -m pytest
