### 2024-04-29 AnkitBarik/planetMagFields#9 @AnkitBarik
I have added a paragraph in the [README](https://github.com/AnkitBarik/planetMagFields) and the [documentation](https://ankitbarik.github.io/planetMagFields/) , please check if this looks good. Thank you!
