### 2024-04-30 AnkitBarik/planetMagFields#11 @athulpg007
Error in plotting example when using Hammer projection
Following the example [here](https://ankitbarik.github.io/planetMagFields/examples.html#planet-plot), attempting to plot the magnetic field using Mollweide projection does not work as expected. To reproduce:
```
In [6]: from planetmagfields import Planet
   ...: p = Planet(name='earth')
   ...: p.plot(r=1,proj='Mollweide')
```
Traceback:
```
Planet: Earth
Model: igrf13
l_max = 13
Dipole tilt (degrees) = -9.983977
Year = 2020
cartopy library not available, using Hammer projection
Cell In[6], line 3
      1 from planetmagfields import Planet
      2 p = Planet(name='earth')
----> 3 p.plot(r=1,proj='Mollweide')

File ~/Documents/planetMagFields03/.venv/lib/python3.11/site-packages/planetmagfields/planet.py:136, in Planet.plot(self, r, levels, cmap, proj)
    133     radLabel = r'  $r/r_{\rm surface}=%.2f$' %r
    135 if proj.lower() != 'hammer' and self.name == 'earth':
--> 136     ax.coastlines()
    138 if r==1:
    139     radLabel = '  Surface'

AttributeError: 'Axes' object has no attribute 'coastlines'
```
Suggestions:
1. If the Mollweide projection is frequently used, consider adding cartopy to the requirements. This error does not occur if Cartopy is installed using pip install cartopy.
2. Please consider fixing the  'Axes' object has no attribute 'coastlines' error when falling back to Hammer projection if cartopy is not available.
