### 2024-04-30 #6677 @athulpg007
Thanks for addressing the above issues. Please also see: https://github.com/AnkitBarik/planetMagFields/issues/11
