### 2024-04-30 23631c9
Fix coastline plotting without cartopy
