### 2024-04-30 AnkitBarik/planetMagFields#11 @AnkitBarik
Thank you for pointing this out. Fixed in 23631c95857d8d53034ee573d0361e9672734445 .
