### 2024-04-30 AnkitBarik/planetMagFields#11 @AnkitBarik
Also bumped patch number on [PyPI](https://pypi.org/project/planetMagFields/).
