+++ .github/workflows/main.yml
@@ -34,7 +34,8 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get -y install python3-scipy python3-numpy python3-matplotlib python3-pytest
-
+          sudo apt-get -y install python3-pip
+          python3 -m pip install pyevtk
       - name: Run test
         run: |
           cd ${{github.workspace}}/tests
