+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "planetMagFields"
-version = "1.4.1"
+version = "1.4.2"
 authors = [
   { name="Ankit Barik", email="ankit.barik@gmail.com" },
 ]
