### 2024-04-30 c5737f2
Fix shtns installation
