### 2024-04-30 28e8101
Fix shtns installation
