+++ .github/workflows/main.yml
@@ -33,10 +33,12 @@ jobs:
       - name: Install pre-requisites
         run: |
           sudo apt-get update
-          sudo apt-get -y install python3-scipy python3-numpy python3-matplotlib python3-pytest
-          sudo apt-get -y install python3-pip
           sudo apt-get -y install libfftw3-dev gcc g++
-          python3 -m pip install pyevtk shtns
+          wget https://bootstrap.pypa.io/get-pip.py
+          python3 get-pip.py
+          python3 -m pip install numpy scipy matplotlib pytest
+          python3 -m pip install pyevtk
+          python3 -m pip install shtns
       - name: Run test
         run: |
           cd ${{github.workspace}}/tests
