### 2024-04-30 ba0f624
Fix test_vtk to account for some size difference
