+++ tests/test_vtk.py
@@ -10,5 +10,5 @@ def test_vtk():
     size = os.path.getsize('earth.vts')
     size_ref = 67109652
     err = abs(size - size_ref)
-    np.testing.assert_allclose(err, 0, rtol=1, atol=1)
+    np.testing.assert_allclose(err, 0, rtol=100, atol=100)
     os.remove('earth.vts')
