### 2024-04-30 AnkitBarik/planetMagFields#10 @AnkitBarik
3. Finished implementing a few different tests:
 - Simple test of the Planet class to reproduce Earth's surface field
 - Two tests for testing potential extrapolation using Jupiter's field
 - Testing data reading and comparing against reference values
 - Testing vts file writing and comparing file size written

The workflow has been updated accordingly.

Hope that works! Thank you.
