### 2024-05-02 AnkitBarik/planetMagFields#10 @athulpg007
Thanks for implementing the tests.
