### 2024-05-02 AnkitBarik/planetMagFields#11 @athulpg007
Thanks for fixing this.
