### 2024-05-02 #6677 @athulpg007
I have completed my initial review of the submission. This is a useful package for researchers working in the area of planetary magnetic fields by providing a fast method of plotting field data for various planets. The package is well written, along with documentation and examples to help new users get started.

My concerns raised earlier have been fixed, thanks. Please see further comments below:

1. According to the JOSS formatting [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#how-should-my-paper-be-formatted), the paper should be between 250-1000 words. Currently, the word count is ~2k. Please consider revising the paper to reduce the length. Some suggestions are given below:

- Consider shortening the "Mathematics" section to give a high-level overview, and move details to the documentation if needed.
- In "Description of Software, Software package", consider removing the input lines [3, 4] which shows some coefficients. The example can be limited to showing how to plot a field.
- Consider shortening the "Jupyter frontend" section, it would suffice to include a link to the binder notebook. Consider removing the line "The utilization of an interactive Jupyter notebook for the exploration of planetary magnetic fields represents a significant advancement in the pedagogical and research approach to planetary magnetic fields.", as the use of an existing general data science tool such as a Jupyter notebook in itself cannot be considered a "significant advancement" in the field of research.
- Consider if the screenshot figures 5 and 6 can be removed, .

My suggestion is to shorten the paper to ~5 pages or less, including references. It would then be easier for a diverse, non-specialist audienceto quickly get a high-level understanding of the package, and then refer to the documentation for more mathematical details, example Jupyter notebook etc.

2. I could run the interactive Jupyter Notebook locally. When trying to run the binder online, the page loaded but the interactive notebook kept trying to load, but could not succeed. Please verify if the binder notebooks still work as expected.

Thanks!
