### 2024-05-02 bfe1fa2
Update doc with math and shorten paper
