+++ paper/paper.md
@@ -27,64 +27,35 @@ tags:
 Long term observations and space missions have generated a wealth of data on the magnetic fields of the Earth and other solar system planets [@IGRF13;@Connerney1987;@Connerney1991;@Kivelson2002;@Anderson2012;@Cao2020;@Connerney2022]. `planetMagfields` is a Python package designed to have all the planetary magnetic field data currently available in one place and to provide an easy interface to access the data. `planetMagfields` focuses on planetary bodies that generate their own magnetic field, namely Mercury, Earth, Jupiter, Saturn, Uranus, Neptune and Ganymede. `planetMagfields` provides functions to compute as well as plot the magnetic field on the planetary surface or at a distance above or under the surface. It also provides functions to filter out the field to large or small scales as well as to produce `.vts` files to visualize the field in 3D using Paraview [@Ahrens2005;@Ayachit2015], VisIt [@visit] or similar rendering software. Lastly, the `planetMagfields` repository also provides a Jupyter notebook for easy interactive visualizations.
 
 # Statement of need
-Planetary scientists studying the magnetic field of planets need to constantly access, visualize, analyze and extrapolate magnetic field data.  In addition, with technological advancements in space exploration and planetary missions, we are constantly getting new data for planetary magnetic fields and hence, better field models. Though reviews of these field models are often written [@Schubert2011;@Stanley2014], there is very little software available that provides easy access to these models with a high level language and a way to easily visualize and analyze them. To the knowledge of the authors, there are a few publicly available repositories that are capable of providing access to planetary magnetic field data and tools to analyze them such as `JupiterMag` [@Wilson2023;@James2024],`KMAG` [@Khurana2020], `ChaosMagPy` [@Kloss2024] and `libinteralfield` (<https://github.com/mattkjames7/libinternalfield>). The first two of these are dedicated towards Jupiter and Saturn, respectively. `ChaosMagPy` is a python interface for the CHAOS model of the Earth's magnetic field [@Finlay2020]. Out of these, only `libinteralfield` provides software to analyze and access magnetic fields of all planets. However, it is a `C++` library which needs to be interfaced with something at a higher level to enable fast analyses and visualization.  Thus, a software package that has different magnetic field models for all different planets of the solar system in one place, as well as provides a high level API to access, analyze and visualize them is not available. `planetMagfields` is intended not only to currently fill this gap, but also to provide a central repository, to be constantly updated, as more magnetic field models become available.
[…]
