+++ doc/bib.bib
@@ -387,20 +387,7 @@ archivePrefix = {arXiv},
   url          = {https://doi.org/10.5281/zenodo.4080294}
 }
 
-@article{Wieczorek2018,
-       author = {{Wieczorek}, Mark A. and {Meschede}, Matthias},
-        title = "{SHTools: Tools for Working with Spherical Harmonics}",
-      journal = {Geochemistry, Geophysics, Geosystems},
-     keywords = {spherical harmonics, software, python, Fortran},
-         year = 2018,
-        month = aug,
-       volume = {19},
-       number = {8},
-        pages = {2574-2592},
-          doi = {10.1029/2018GC007529},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2018GGG....19.2574W},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+
 
 % Python packages
 
