### 2024-05-02 #6677 @AnkitBarik
@athulpg007 Thank you very much for the comments. I have made the following changes:

 1.
     - Added an additional reference I had missed earlier to the package [SHTools](https://shtools.github.io/SHTOOLS/index.html).
     - Shortened the "Mathematics" section and removed the figure on there and moved the details to the [documentation](https://ankitbarik.github.io/planetMagFields/math.html). I think the remaining description is required for readers to understand what is going on in the package. If you think it's still too much, I can try cutting it down even more with a link to the documentation.
     - Removed input lines [3,4] as suggested from "Description of Software, Software package".
     - Shortened the "Jupyter frontend" section significantly with only a sentence and the link to the binder.
     - Removed screenshot figures 5 and 6.
   The main text of the paper is now 4 pages long (including figures). However, because of the large number of references, the total length is 7 pages. I believe the references should be there as the libraries and other studies used, need to be cited. If you have any other suggestions for slashing down on content, I would be happy to do so.

2. I verified on my end that the binder works as intended. I'm not entirely sure why it didn't work.

Thank you!
