### 2024-05-03 #6677 @kjg136
I have completed my review of the submission, including the changes made in response to the first reviewer's comments. I agree that the package is useful for visualizing and analyzing planetary magnetic fields. To my knowledge, the authors are correct in their statement that no such package for a high-level language is currently established and therefore, the submission represents a valuable contribution. The documentation and examples are clear, making the package easy to understand and use.

I see no major issues and wish to only make two quick comments:

1. Grammatical error in line 53 of the paper: ...as how that field looks ~~like~~ at...
2. I can confirm I was able to use the binder.
