+++ paper/paper.md
@@ -42,7 +42,7 @@ Magnetic fields in planets are generated by electric currents in a fluid region 
 
 where, $g_l^m$ and $h_l^m$ are called the Gauss coefficients. $R_p$ represents the radius of the planet and $P_l^m$ are associated Legendre functions of order $l$ and degree $m$, where $l$ and $m$ are integers. The above equation can be recast in terms of spherical harmonics, which is what the code uses.
 
-The raw data obtained from satellites or space missions are usually inverted to obtain these Gauss coefficients which are the key to describing the surface magnetic field of a planet as well as how that field looks like at a certain altitude from the surface. The magnetic energy content on the surface in a certain degree $l$ is given by the Lowes spectrum:
+The raw data obtained from satellites or space missions are usually inverted to obtain these Gauss coefficients which are the key to describing the surface magnetic field of a planet as well as how that field looks at a certain altitude from the surface. The magnetic energy content on the surface in a certain degree $l$ is given by the Lowes spectrum:
 
 $$R_{l} = (l + 1) \sum_{m}\left( \left(g_l^m\right)^2 + \left(h_l^m\right)^2\right),$$
 
