### 2024-05-05 aac37bd
Update planetMag citation
