+++ paper/paper.bib
@@ -402,6 +402,18 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@software{Styczinski2024,
+  author       = {{Styczinski}, Marshall J. and {Cochrane}, Corey J.},
+  title        = {{coreyjcochrane/PlanetMag: Model updates following
+                   publication peer review}},
+  month        = mar,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v1.0.2},
+  doi          = {10.5281/zenodo.10864719},
+  url          = {https://doi.org/10.5281/zenodo.10864719}
+}
+
 % Python packages
 
 @article{ipython,
