### 2024-05-05 #6677 @AnkitBarik
I made one last change to the paper, I updated the citation for `PlanetMag` to the correct [zenodo doi](https://zenodo.org/records/10864719).
