### 2024-05-05 9c6fcbb
Tiny edits for JOSS publication
