### 2024-05-05 #6677 @dfm
@AnkitBarik — I've opened a small PR with some minor edits to the manuscript, please take a look and merge or let me know what you think.

Once you've done that:
