+++ doc/conf.py
@@ -14,7 +14,7 @@ sys.path.insert(0, os.path.abspath('.'))
 project = 'planetMagFields'
 copyright = '2024, Ankit Barik'
 author = 'Ankit Barik'
-release = '1.4'
+release = '1.4.3'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
