### 2024-05-06 #6677 @AnkitBarik
@dfm Thank you very much for the minor edits.

 1. The manuscript seems okay to me.
 2. I have bumped up the version number to **1.4.3** everywhere (including PyPI).
 3. A zenodo release has been automatically created through GitHub : https://zenodo.org/doi/10.5281/zenodo.4690524 .
 *However*, because of your pull request, it seems like your name got added to the author list. Not sure if there's a fix for that.
