+++ paper/paper.bib
@@ -21,7 +21,7 @@ pages = {93-105},
 year = {2018},
 note = {Special Issue:15th SEDI conference},
 issn = {0031-9201},
-doi = {https://doi.org/10.1016/j.pepi.2017.07.001},
+doi = {10.1016/j.pepi.2017.07.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0031920117300304},
 author = {Thébault, E.  and Langlais, B. and Oliveira, J.S. and Amit, H. and Leclercq, L.},
 }
@@ -171,7 +171,7 @@ number = {14},
 pages = {1706-1713},
 year = {2009},
 issn = {0032-0633},
-doi = {https://doi.org/10.1016/j.pss.2009.04.008},
+doi = {10.1016/j.pss.2009.04.008},
 url = {https://www.sciencedirect.com/science/article/pii/S003206330900107X},
 author = {M.E. Burton and M.K. Dougherty and C.T. Russell},
 keywords = {Saturn magnetic field, Magnetosphere, Planetary magnetic fields},
@@ -201,7 +201,7 @@ pages = {113541},
 year = {2020},
 note = {Cassini Mission Science Results},
 issn = {0019-1035},
-doi = {https://doi.org/10.1016/j.icarus.2019.113541},
+doi = {10.1016/j.icarus.2019.113541},
 url = {https://www.sciencedirect.com/science/article/pii/S0019103519305299},
 author = {Hao Cao and Michele K. Dougherty and Gregory J. Hunt and Gabrielle Provan and Stanley W.H. Cowley and Emma J. Bunce and Stephen Kellock and David J. Stevenson},
 keywords = {Saturn, Magnetic fields, Geophysics, Interior},
@@ -231,7 +231,7 @@ address = {Boston},
 pages = {121-136},
 year = {2014},
 isbn = {978-0-12-415845-0},
-doi = {https://doi.org/10.1016/B978-0-12-415845-0.00006-2},
+doi = {10.1016/B978-0-12-415845-0.00006-2},
 url = {https://www.sciencedirect.com/science/article/pii/B9780124158450000062},
 author = {Sabine Stanley},
 }
@@ -272,7 +272,7 @@ address = {Burlington},
 pages = {717-731},
 year = {2005},
 isbn = {978-0-12-387582-2},
-doi = {https://doi.org/10.1016/B978-012387582-2/50038-1},
+doi = {10.1016/B978-012387582-2/50038-1},
 url = {https://www.sciencedirect.com/science/article/pii/B9780123875822500381},
 author = {Ahrens, James and Geveci, Berk and Law, Charles}
 }
