### 2024-05-09 #6677 @dfm
@AnkitBarik — I know that that error message is a little opaque, but my guess is that the issue here is with the use of `dfrac` in the manuscript. Could you try switching it to `frac` and we can try this again?
