+++ paper/paper.md
@@ -36,7 +36,7 @@ In addition to the research aspect of our software, the interactive Jupyter note
 Magnetic fields in planets are generated by electric currents in a fluid region inside them through a process called dynamo action [@Jones2011;@Schubert2011;@Stanley2014]. Outside this region, in the absence of current sources, the magnetic field $\vec{B}$ can be written as the gradient of a scalar potential, $\vec{B} = -\nabla V$. The potential $V$ is usually written as an expansion in orthogonal functions in spherical coordinates $(r,\theta,\phi)$,
 
 \begin{equation}
-  V = R_p \sum_{l,m} \left(\dfrac{R_p}{r}\right)^{l+1} [g_l^m \cos(m\phi) + h_l^m \sin(m\phi)] P_l^m (\cos\theta)\, ,
+  V = R_p \sum_{l,m} \left(\frac{R_p}{r}\right)^{l+1} [g_l^m \cos(m\phi) + h_l^m \sin(m\phi)] P_l^m (\cos\theta)\, ,
   \label{eq:gaussCoeff}
 \end{equation}
 
