### 2024-05-09 #6677 @AnkitBarik
@dfm I have switched it to \frac and verified that it still looks fine on the pdf.
