### 2024-05-09 #6677 @AnkitBarik
@dfm @athulpg007 @kjg136 This is wonderful! Thank you very much for help and reviews ! 🎉🎉🎉
@dfm I had one question : what is JOSS policy for cross-posting on Arxiv?
