### 2024-05-10 #6677 @dfm
> @dfm I had one question : what is JOSS policy for cross-posting on Arxiv?

JOSS is happy for you to post to arXiv and you can use `@editorialbot generate preprint` to produce the required `.tex` file. Sometimes arXiv rejects JOSS papers as out of scope and there's not too much we can do about that unfortunately.
